MVP-002: extract overview and history tabs from PatientCardPage
